Added via upload explanation:
Ran into difficulties adding via command line with token use and login authentication. Included link to video showing I created these things and completed the assignment on local command line but due to time constraints (this week, and the last three have sucked) I may not be able to rectify the issue by the submission deadline and hope to obtain some points in the event that I run out of time and this is my submission.